+++ ThirdPartyAdapters/mintegral/mintegral/src/test/java/com/google/ads/mediation/mintegral/MintegralMediationAdapterTest.kt
@@ -23,6 +23,7 @@ import com.google.ads.mediation.adaptertestkit.AdapterTestKitConstants.TEST_WATE
 import com.google.ads.mediation.adaptertestkit.assertGetSdkVersion
 import com.google.ads.mediation.adaptertestkit.assertGetVersionInfo
 import com.google.ads.mediation.adaptertestkit.createMediationAppOpenAdConfiguration
+import com.google.ads.mediation.adaptertestkit.createMediationBannerAdConfiguration
 import com.google.ads.mediation.adaptertestkit.createMediationConfiguration
 import com.google.ads.mediation.adaptertestkit.createMediationInterstitialAdConfiguration
 import com.google.ads.mediation.adaptertestkit.loadAppOpenAdWithFailure
@@ -30,19 +31,26 @@ import com.google.ads.mediation.adaptertestkit.loadInterstitialAdWithFailure
 import com.google.ads.mediation.adaptertestkit.loadRtbAppOpenAdWithFailure
 import com.google.ads.mediation.adaptertestkit.loadRtbInterstitialAdWithFailure
 import com.google.ads.mediation.mintegral.MintegralConstants.AD_UNIT_ID
+import com.google.ads.mediation.mintegral.MintegralConstants.ERROR_BANNER_SIZE_UNSUPPORTED
+import com.google.ads.mediation.mintegral.MintegralConstants.ERROR_DOMAIN
 import com.google.ads.mediation.mintegral.MintegralConstants.PLACEMENT_ID
 import com.google.ads.mediation.mintegral.MintegralUtils.getAdapterVersion
 import com.google.ads.mediation.mintegral.MintegralUtils.getSdkVersion
 import com.google.android.gms.ads.AdError
 import com.google.android.gms.ads.AdFormat
+import com.google.android.gms.ads.AdSize
 import com.google.android.gms.ads.MobileAds
 import com.google.android.gms.ads.RequestConfiguration
 import com.google.android.gms.ads.mediation.InitializationCompleteCallback
 import com.google.android.gms.ads.mediation.MediationAdLoadCallback
 import com.google.android.gms.ads.mediation.MediationAppOpenAd
 import com.google.android.gms.ads.mediation.MediationAppOpenAdCallback
+import com.google.android.gms.ads.mediation.MediationBannerAd
+import com.google.android.gms.ads.mediation.MediationBannerAdCallback
 import com.google.android.gms.ads.mediation.MediationInterstitialAd
 import com.google.android.gms.ads.mediation.MediationInterstitialAdCallback
+import com.google.common.truth.Truth.assertThat
+import com.mbridge.msdk.out.MBBannerView
 import com.mbridge.msdk.out.MBridgeSDKFactory
 import com.mbridge.msdk.out.SDKInitStatusListener
 import com.mbridge.msdk.system.MBridgeSDKImpl
@@ -77,6 +85,10 @@ class MintegralMediationAdapterTest {
   private val mockAppOpenAdLoadCallback:
     MediationAdLoadCallback<MediationAppOpenAd, MediationAppOpenAdCallback> =
     mock()
+  private val mediationBannerAdLoadCallback:
+    MediationAdLoadCallback<MediationBannerAd, MediationBannerAdCallback> =
+    mock()
+  private val mbBannerView: MBBannerView = mock()
 
   @Before
   fun setUp() {
@@ -162,6 +174,54 @@ class MintegralMediationAdapterTest {
       initStatusCaptor.firstValue.onInitSuccess()
       verify(mockMBridgeSdk).setCoppaStatus(eq(context), eq(true))
       verify(mockInitializationCompleteCallback).onInitializationSucceeded()
+    }
+  }
+
+  // endregion
+
+  // region Banner Ad Tests
+
+  @Test
+  fun loadWaterfallBannerAd_forUnsupportedAdSize_fails() {
+    val mediationBannerAdConfiguration =
+      createMediationBannerAdConfiguration(context = context, adSize = AdSize(800, 50))
+
+    mintegralMediationAdapter.loadBannerAd(
+      mediationBannerAdConfiguration,
+      mediationBannerAdLoadCallback,
+    )
+
+    val adErrorCaptor = argumentCaptor<AdError>()
+    verify(mediationBannerAdLoadCallback).onFailure(adErrorCaptor.capture())
+    val adError = adErrorCaptor.firstValue
+    assertThat(adError.code).isEqualTo(ERROR_BANNER_SIZE_UNSUPPORTED)
+    assertThat(adError.domain).isEqualTo(ERROR_DOMAIN)
+  }
+
+  /**
+   * Tests that RTB Banner ad load succeeds even if the requested size is a non-standard one for
+   * Mintegral.
+   */
+  @Test
+  fun loadRtbBannerAd_forNonStandardMintegralSize_succeeds() {
+    val mediationBannerAdConfiguration =
+      createMediationBannerAdConfiguration(
+        context = context,
+        serverParameters = bundleOf(AD_UNIT_ID to TEST_AD_UNIT, PLACEMENT_ID to TEST_PLACEMENT_ID),
+        bidResponse = TEST_BID_RESPONSE,
+        watermark = TEST_WATERMARK,
+        adSize = AdSize(800, 50),
+      )
+
+    mockStatic(MintegralFactory::class.java).use {
+      whenever(MintegralFactory.createMBBannerView(any())) doReturn mbBannerView
+
+      mintegralMediationAdapter.loadRtbBannerAd(
+        mediationBannerAdConfiguration,
+        mediationBannerAdLoadCallback,
+      )
+
+      verify(mbBannerView).loadFromBid(TEST_BID_RESPONSE)
     }
   }